Reduce search space by keeping person at a fixed position.
Because the table is a circle, running *all* permutations means
creating a large number of equivalent seatings, just shifted.  By
keeping one person in a fixed position, we can imagine virtually
rotating the table and creating all of those equivalent adjacencies,
but without having to compute the happiness cost of each.